# swagger.yaml
swagger: '2.0'
info:
  title: RiftBattle API
  description: API for managing users, products, orders, Fortnite stats, and more.
  version: 1.0.0
basePath: /api
...
paths:
  /get_link:
    get:
      tags:
        - fortnite
      summary: Get Fortnite authentication link
      responses:
        '200':
          description: Successful response with device code
          schema:
            type: object
            properties:
              user_code:
                type: string
                example: "ABCD-EFGH"
              device_code:
                type: string
                example: "1234567890"
              verification_uri:
                type: string
                example: "https://example.com/verify"
              expires_in:
                type: integer
                example: 600
        '400':
          description: Bad request (e.g., missing or invalid parameters)
          schema:
            type: object
            properties:
              error:
                type: string
                example: "device_code is required"
        '500':
          description: Internal server error
          schema:
            type: object
            properties:
              error:
                type: string
                example: "internal server error"
...
  /get_role:
    get:
      tags:
        - user
      summary: Get the role of the current user
      responses:
        '200':
          description: Successful response with user role
          schema:
            type: string
            example: "user"
        '500':
          description: Internal server error
          schema:
            type: object
            properties:
              error:
                type: string
                example: "internal server error"
  /login:
    post:
      tags:
        - user
      summary: Create a session for the user
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
                example: "user@example.com"
              code:
                type: integer
                example: 123456
      responses:
        '200':
          description: Session created successfully
        '400':
          description: Bad request (e.g., invalid OTP)
          schema:
            type: object
            properties:
              error:
                type: string
                example: "invalid OTP"
        '401':
          description: Unauthorized (e.g., invalid credentials)
          schema:
            type: object
            properties:
              error:
                type: string
                example: "unauthorized"
        '500':
          description: Internal server error
          schema:
            type: object
            properties:
              error:
                type: string
                example: "internal server error"
...
  /create_order:
    post:
      tags:
        - order
      summary: Create a new order
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
                example: "user@example.com"
              product_id:
                type: integer
                example: 1
              payment_method:
                type: string
                example: "balance"
              coupon:
                type: string
                example: "DISCOUNT10"
              warranty:
                type: number
                format: float
                example: 3
      responses:
        '200':
          description: Order created successfully
          schema:
            $ref: '#/definitions/Order'
        '400':
          description: Bad request (e.g., invalid product or payment method)
          schema:
            type: object
            properties:
              error:
                type: string
                example: "invalid payment method"
        '500':
          description: Internal server error
          schema:
            type: object
            properties:
              error:
                type: string
                example: "internal server error"
...
  /get_product:
    get:
      tags:
        - product
      summary: Get a specific product by ID
      parameters:
        - name: id
          in: query
          type: integer
          required: true
          description: Product ID
      responses:
        '200':
          description: Successful response with product details
          schema:
            $ref: '#/definitions/Product'
        '400':
          description: Bad request (e.g., invalid product ID)
          schema:
            type: object
            properties:
              error:
                type: string
                example: "invalid product ID"
        '500':
          description: Internal server error
          schema:
            type: object
            properties:
              error:
                type: string
                example: "internal server error"
...
definitions:
...
  Order:
    type: object
    properties:
      id:
        type: integer
        example: 1
      email:
        type: string
        example: "user@example.com"
      total_price:
        type: number
        format: float
        example: 50.0
      user_id:
        type: integer
        example: 1
      product_id:
        type: integer
        example: 1
      seller_id:
        type: integer
        example: 2
      payment_method:
        type: string
        example: "balance"
      coupon:
        type: string
        example: "DISCOUNT10"
      status:
        type: string
        example: "pending"
      warranty:
        type: integer
        example: 3
      date:
        type: integer
        format: int64
        example: 1690924800
...
  Product:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: "Fortnite Account"
      price:
        type: number
        format: float
        example: 50.0
      credentials:
        type: string
        example: "username:password"
      seller_id:
        type: integer
        example: 2
      athena_ids:
        type: array
        items:
          type: string
        example: ["CID_123", "CID_456"]
      stats:
        type: object
        example: {"level": "100", "backpacks": "500"}
      description:
        type: string
        example: "A rare Fortnite account"
      date:
        type: integer
        format: int64
        example: 1690924800
      active:
        type: boolean
        example: true
      discount:
        type: integer
        example: 10
...